stdf-data-factory/requirements.txt
streamlit
pymupdf
markitdown
groq
openai
pillow
pandas
python-dotenv
plotly
crawl4ai